# Minor changes to speed up asynchronous label propagation for community detection.