@dschult 2022-01-06
That makes good sense.  It is faster (even for the cases where counts are weighted) to use an iterator rather than a loop over `update` calls, or even a single update over a constructed dict (or list).  We want to both reduce the calls to `update` and reduce creating dicts or lists which just get sent to `update`.

How about this:
```python
if weight is None:
    edge_info = iter(G[node])
else:
    edge_info = ((v, edge_wt) for _, v, edge_wt in G.edges(node, data=weight, default=1))
label_freq = Counter()
label_freq.update(edge_info)
```
I get this as much faster for both weighted and unweighted cases.  Do you get it faster for the unweighted case?
And more importantly, did I do it right? Are we still getting the same answers?