@leojklarner 2022-01-07
Thank you for the reply.

Using an iterator instead of an explicitly created list is a good idea. However, the counter needs to use the nodes' current labels as keys, rather than the node IDs, so you'd need to add an additional call to the labels dict for each neighbour.

Using an iterator for the unweighted case gives an additional 25% speed-up over the list comprehension:
```
edge_info = map(labels.get, G[node])
```
however the [Counter's update method](https://docs.python.org/3/library/collections.html#collections.Counter) does not accept iterables of key-value pairs, so
```
edge_info = ((label[v], wt) for _, v, wt in G.edges(node, data=weight, default=1))
label_freq.update(edge_info)
```
would just count the frequencies of (label, edge_weight) pairs, while
```
edge_info = {label[v]: wt for _, v, wt in G.edges(node, data=weight, default=1)}
label_freq.update(edge_info)
```
would just return the edge weight of the last occurrence of each label.

So given my understanding that the edge weights should be used as multipliers for the count of the corresponding label, it seems that either an incremental for-loop or a list with `wt` repetitions of the corresponding label is needed, of which the former is preferable in my opinion.