@leojklarner 2022-01-11
I implemented the defaultdict approach mentioned above. Thank you for the feedback and taking the time to review!